+++ R/add_tooltip.R
@@ -75,3 +75,4 @@ add_tooltip <-
 
     nodes_and_edges
   }
+