+++ NEWS.md
@@ -1,5 +1,7 @@
 # chariotViz 0.0.2.9000  
 
+## New Features  
+
 * Added `omop.ancestors` class objects for taxonomy visualization  
 * Begin developing features for ancestor visualization  
 * Rename `fetch_omop` function to `fetch_omop_relationships` and `fetch_omop_ancestors` 
@@ -7,6 +9,9 @@ to accomodate new ancestors feature
 * Added `filter_for_*_relationships` and `filter_out_*_relationships` functions to 
 refine omop relationships visualization  
 
+## Bugs  
+
+* Remove printing of column names when calling `chariotViz()`